requirements_pdf.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0 